+++ lib/tvrage_api/tvrage_api.py
@@ -22,7 +22,6 @@ import datetime as dt
 import requests
 import requests.exceptions
 import xmltodict
-from sickbeard.network_timezones import standardize_network
 
 try:
     import xml.etree.cElementTree as ElementTree
@@ -444,9 +443,7 @@ class TVRage:
             if value:
                 if isinstance(value, dict):
                     if key == 'network':
-                        network = value['#text']
-                        country = value['@country']
-                        value = standardize_network(network, country)
+                        value = value['#text']
                     if key == 'genre':
                         value = value['genre']
                         if not value: